pyproject.toml
[project]
name = "frotzmark"
version = "1.0.0"
description = "LLMs vs Interactive Fiction: A framework for evaluating language models on text adventure games"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.0.1",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
logfire = [
    "logfire>=2.6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
